+++ ide/h1st/src/browser/visualization/containers/GrapphContainer.tsx
@@ -1,0 +1,13 @@
+import * as React from "react";
+import { Provider } from "react-redux";
+import Graph from "../components/graph";
+
+export const GraphContainer = (props: { store: any }) => {
+  return (
+    <div style={{ height: "100%" }}>
+      <Provider store={props.store}>
+        <Graph />
+      </Provider>
+    </div>
+  );
+};
