+++ ide/h1st/src/browser/visualization/reducers/index.ts
@@ -1,0 +1,4 @@
+import { combineReducers } from "redux";
+import graph from "./graph";
+
+export default combineReducers({ graph });
