+++ ide/h1st/src/browser/visualization/types/graph.ts
@@ -1,0 +1,23 @@
+type NodeType = "condition" | "action" | undefined;
+
+export interface INode {
+  id?: string;
+  name?: string;
+  type: NodeType | string;
+  subtype?: string;
+  x: number;
+  y: number;
+  title?: string;
+}
+
+export interface IEdge {
+  handleText: string;
+  source: string;
+  target: string;
+  type: string;
+}
+
+export interface IGraph {
+  nodes: INode[];
+  edges: IEdge[];
+}
