+++ ide/h1st/src/browser/visualization/types/index.ts
@@ -1,0 +1,6 @@
+import { IGraph } from "./graph";
+
+export interface IVisualizationStore {
+  graph: IGraph;
+}
+export * from "./graph";
